Added Canvas Scale Broadcaster
Send SetDirty to all components when the screen is scaling.
This resolves the problem that there is no such event in the UI system.

You can change or add messages being sent on scaling.

It's a quick hack.
Performance might be terrible, as it uses a BroadcastMessage to send the
message out.